requirements.txt
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.6
feedparser==6.0.10
idna==3.4
joblib==1.2.0
lxml==4.9.2
numpy==1.24.1
pandas==1.5.3
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
scikit-learn==1.2.0
scipy==1.10.0
sgmllib3k==1.0.0
six==1.16.0
sklearn==0.0.post1
soupsieve==2.4
threadpoolctl==3.1.0
tqdm==4.64.1
twilio==7.16.4
urllib3==1.26.14
